Refactor: Code structure corrections and README improvements
