+++ MicrobialAntigenPipline/run.sh
@@ -1,0 +1,3 @@
+export TF_CPP_MIN_LOG_LEVEL=2. 
+export CUDA_VISIBLE_DEVICES="".
+python MicrobialAntigen.py -c ../../../configures/MicrobialAntigen_configure.yaml -p ../../../configures/MicrobialAntigen_pathes.yaml
