--- PiplineOfAntigenDetection/MicrobialAntigenPipline/configures/MicrobialAntigen_configure.yaml
@@ -1,36 +1,0 @@
-# Configuration Paths
-path:
- tmp_dir : "<project_root>/tmp/"  # Temporary directory for processing files
- input_dir : "<project_root>/input_data/"  # Input data directory
- output_dir : "<project_root>/analysis_output/"  # Output directory
-
-# Runtime Parameters
-args:
- thread : 20  # CPU threads allocated per sample
- pool_size : 2  # Number of samples processed concurrently
- mem : "128G"  # Maximum memory allocation
- mem_perthread : "256M"  # Memory per thread in samtools sort
-
-# Analysis Parameters
-others:
- species : "human"  # Species selection (human/mouse)
- seq_type : "rna"  # Sequencing type (wgs/rna)
- pair : True  # Paired-end sequencing status
- QC : True  # Enable quality control steps
- host_sequences_removing : True  # Enable host sequence removal
- microbial_taxas_quantification : True  # Enable microbial taxonomy analysis
- match_length_threshold : 0.95  # Alignment length threshold
- score_threshold : 1  # Minimum alignment score
- microbial_peptides_identification : True  # Enable peptide identification
- hlatyping : True  # Enable HLA typing
- time_out : 36000  # Process timeout in seconds
- microbial_peptides_bindingPrediction : True  # Enable binding prediction
-
-# Sample List (Note: Numeric sample IDs require quotes)
-samples:
-- Sample_1
-- Sample_2
-- Sample_3
-- Sample_4
-- Sample_5
-- Sample_6
