+++ configures/MicrobialAntigen_pathes.yaml
@@ -1,0 +1,49 @@
+# Resource Path Configuration
+path:
+  # Host reference genomes
+  host_fa_hg38: '/path/to/GRCh38/GCA_000001405.15_GRCh38_no_alt_analysis_set.fa'
+  host_img_hg38: '/path/to/GRCh38/GCA_000001405.15_GRCh38_no_alt_analysis_set.fa.img'
+  host_kmer_hg38: '/path/to/GRCh38/GCA_000001405.15_GRCh38_no_alt_analysis_set.hss'
+
+  host_fa_t2t: '/path/to/t2t/chm13v2.0.fa'
+  host_img_t2t: '/path/to/t2t/chm13v2.0.fa.img'
+  host_kmer_t2t: '/path/to/t2t/chm13v2.0.hss'
+  
+  host_fa_GRCm38: '/path/to/pathseq/GRCm38/GRCm38.fa'
+  host_img_GRCm38: '/path/to/pathseq/GRCm38/GRCm38.fa.img'
+  host_kmer_GRCm38: '/path/to/pathseq/GRCm38/GRCm38.hss'
+  
+  # Microbial database (preserve last 3 components)
+  microbe_dict: '/path/to/microbiome_v7/pathseq_custom_V7_fixed.dict'
+  microbe_img: '/path/to/microbiome_v7/pathseq_custom_V7_fixed.fa.img'
+  taxonomy_file: '/path/to/microbiome_v7/pathseq_custom_V7_fixed.db'
+  ncid2microbeNames_file: '/path/to/microbiome_v7/pathseq_custom_V7_fixed_microbeNames_sep=tab.csv'
+  microbial_genome_length: '/path/to/microbiome_v7/pathseq_custom_V7_fixed_GenomeLength.csv'
+  tax_id_hierarchy: '/path/to/microbiome_v7/V7_taxid_hierarchy.txt'
+  catalog_genome: '/path/to/microbiome_v7/V7_catalog_genome.txt'
+  
+  # Tool paths (preserve last 3 components)
+  java17: '/opt/conda/envs/java17/bin/java'
+  gatk_4.6_jar: '/opt/biotools/gatk-4.6.0.0/gatk-package-4.6.0.0-local.jar'
+  fastp: '/opt/conda/envs/RNA-seq/bin/fastp'
+
+  # HLA typing resources
+  freq_data_dir: '/path/to/hlahd.1.7.0/freq_data/'
+  HLA_gene: '/path/to/hlahd.1.7.0/HLA_gene.split.txt'
+  dictionary: '/path/to/hlahd.1.7.0/dictionary/'
+  picard_path: "/path/to/hlahd1.7.0_new/picard.jar"
+  hla_gen: "/path/to/hla_gen/hla_gen"
+
+  # pVacTools configuration
+  sif_file: '/opt/containers/pvactools.4.0.3.sif'
+  pvacbind: '/opt/conda/envs/pvactools/bin/pvacbind'
+  iedb-install-directory: '/opt/db/IEDB/'
+
+  # R
+  R_HOME: '/opt/conda/envs/r4.3/lib/R/'
+  R_LIBRARY: "/opt/conda/envs/r4.3/lib/R/library"
+
+# Database Configuration
+database:
+  db_dir: "/path/to/custom_protein_V3"  # Blast database path
+  outfmt: "\"6 qseqid sseqid qseq sseq stitle pident length mismatch gapopen evalue bitscore\""  # blastx output format
